refactor(structure): create folder for each component
References:
- https://github.com/rascui/website/issues/28
