+++ website-react/src/App.js
@@ -1,14 +1,14 @@
 import './assets/css/style.css';
-import PageOverlay from './components/PageOverlay';
-import PageHeader from './components/PageHeader';
-import PageFooter from './components/PageFooter';
-import PageProgress from './components/PageProgress';
-import SectionHome from './components/SectionHome';
-import SectionAbout from './components/SectionAbout';
-import SectionServices from './components/SectionServices';
-import SectionPortfolio from './components/SectionPortfolio';
-import SectionBlog from './components/SectionBlog';
-import SectionStats from './components/SectionStats';
+import PageOverlay from './components/PageOverlay/PageOverlay';
+import PageHeader from './components/PageHeader/PageHeader';
+import PageFooter from './components/PageFooter/PageFooter';
+import PageProgress from './components/PageProgress/PageProgress';
+import SectionHome from './components/SectionHome/SectionHome';
+import SectionAbout from './components/SectionAbout/SectionAbout';
+import SectionServices from './components/SectionServices/SectionServices';
+import SectionPortfolio from './components/SectionPortfolio/SectionPortfolio';
+import SectionBlog from './components/SectionBlog/SectionBlog';
+import SectionStats from './components/SectionStats/SectionStats';
 
 function App() {
   return (
