+++ website-react/src/components/PageHeader/PageHeader.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import ImageLogo from '../assets/images/O-portfolio/pictures/IEEE Text-1.png';
+import ImageLogo from '../../assets/images/O-portfolio/pictures/IEEE Text-1.png';
 
 const PageHeader = () => {
 	const [menuOpen, setMenuOpen] = useState(false);
